+++ src/Pim/Bundle/BaseConnectorBundle/Tests/Unit/Reader/File/CsvReaderTest.php
@@ -18,8 +18,7 @@ class CsvReaderTest extends \PHPUnit_Framework_TestCase
      */
     protected function setUp()
     {
-        $this->archiver = $this->getArchiverMock();
-        $this->reader = new CsvReader($this->archiver);
+        $this->reader = new CsvReader();
         $this->stepExecution = $this->getStepExecutionMock();
         $this->reader->setStepExecution($this->stepExecution);
 
@@ -43,11 +42,6 @@ class CsvReaderTest extends \PHPUnit_Framework_TestCase
     public function testRead()
     {
         $this->reader->setFilePath(__DIR__ . '/../../../fixtures/import.csv');
-
-        $this->archiver
-            ->expects($this->once())
-            ->method('setHeader')
-            ->with(array('firstname', 'lastname', 'age'));
 
         $this->stepExecution
             ->expects($this->exactly(3))
@@ -111,13 +105,6 @@ class CsvReaderTest extends \PHPUnit_Framework_TestCase
     {
         return $this
             ->getMockBuilder('Akeneo\Bundle\BatchBundle\Entity\StepExecution')
-            ->disableOriginalConstructor()
-            ->getMock();
-    }
-
-    protected function getArchiverMock()
-    {
-        return $this->getMockBuilder('Pim\Bundle\BaseConnectorBundle\Archiver\InvalidItemsCsvArchiver')
             ->disableOriginalConstructor()
             ->getMock();
     }